refactor(app): rename ipc to types
